requirements.txt
pyyaml
opencv-python
pandas
tqdm
matplotlib
flow-vis
librosa
transformers==4.30.2
albumentations==1.4.15
albucore==0.0.16
torchdiffeq==0.2.5
timm==1.0.9
face_alignment==1.4.1
av==12.0.0
tensorboard
dlib
moviepy
basicsr
gfpgan==1.3.8
tb-nightly #==2.13.0a20230423 #-i https://mirrors.aliyun.com/pypi/simple
facexlib
omegaconf
funasr
torchaudio